Renamed binding to networkcamera
