+++ addons/binding/org.openhab.binding.networkcameramotiondetection/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>org.openhab.binding.networkcameramotiondetection</name>
+	<name>org.openhab.binding.networkcamera</name>
 	<comment></comment>
 	<projects>
 	</projects>
