+++ addons/binding/org.openhab.binding.networkcameramotiondetection/ESH-INF/thing/thing-types.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<thing:thing-descriptions bindingId="networkcameramotiondetection" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+<thing:thing-descriptions bindingId="networkcamera" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xmlns:thing="http://eclipse.org/smarthome/schemas/thing-description/v1.0.0"
 	xsi:schemaLocation="http://eclipse.org/smarthome/schemas/thing-description/v1.0.0 http://eclipse.org/smarthome/schemas/thing-description-1.0.0.xsd">
 
-	<thing-type id="networkcamera">
-		<label>Network Camera Motion Detection</label>
+	<thing-type id="motiondetection">
+		<label>Motion Detection</label>
 		<description>Receive images from network camera.</description>
 
 		<channels>
