+++ addons/binding/org.openhab.binding.networkcameramotiondetection/src/main/java/org/openhab/binding/networkcamera/internal/ftp/FtpServerEventListener.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.networkcameramotiondetection.internal.ftp;
+package org.openhab.binding.networkcamera.internal.ftp;
 
 /**
  * This interface defines interface to receive data from FTP server.
